Handle some edge cases when index is incomplete
For example, when a command like 'exit' is issued, an exitcode
will never be recorded, but that leaves an invalid line. We now
assume if no exit was ever given, that it was a exit code 0

Similarly, when a command is squelched via #shh tag, log result
does not know it was squelched, and writes the exit code for the
sequelched command, giving an invalid index.

We now check to make sure the last line in the index is valid
before writing a result onto the end of it to prevent this error.